feat: 인터페이스 명 변경 및 동호회원 전체 조회 시 response에 티어 추가
