--- badminton-api/src/main/java/org/badminton/api/club/validation/ClubNameValid.java
@@ -1,23 +1,0 @@
-package org.badminton.api.club.validation;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-import jakarta.validation.Constraint;
-import jakarta.validation.Payload;
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.Size;
-
-@Constraint(validatedBy = ClubNameValidator.class)
-@Target({ElementType.FIELD, ElementType.PARAMETER})
-@Retention(RetentionPolicy.RUNTIME)
-public @interface ClubNameValid {
-
-	String message() default "동호회 이름은 필수이며 2자 이상 20자 이하여야 합니다";
-
-	Class<?>[] groups() default {};
-
-	Class<? extends Payload>[] payload() default {};
-}
